+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please add your functional changes to the appropriate section in the PR.
 Keep it human-readable, your future self will thank you!
 
-## [Unreleased](https://github.com/ecmwf/anemoi-graphs/compare/0.4.1...HEAD)
+## [Unreleased](https://github.com/ecmwf/anemoi-graphs/compare/0.4.2...HEAD)
+
+## [0.4.2 - Optimisations and lat-lon](https://github.com/ecmwf/anemoi-graphs/compare/0.4.1...0.4.2) - 2024-12-19
 
 ### Added
 
@@ -18,6 +20,7 @@ Keep it human-readable, your future self will thank you!
 - feat: new transform functions to map from sin&cos values to latlon (#98)
 
 ### Changed
+
 - fix: faster edge builder for tri icosahedron. (#92)
 
 ## [0.4.1 - ICON graphs, multiple edge builders and post processors](https://github.com/ecmwf/anemoi-graphs/compare/0.4.0...0.4.1) - 2024-11-26